@@ -1,10 +1,16 @@
 module gofy
 
-go 1.21.0
+go 1.22.3
 
 require (
-	github.com/google/uuid v1.3.1 // indirect
-	github.com/gorilla/mux v1.8.0 // indirect
-	github.com/lib/pq v1.10.9 // indirect
-	golang.org/x/crypto v0.14.0 // indirect
+	github.com/lib/pq v1.10.9
+	github.com/pascaldekloe/jwt v1.12.0
+	golang.org/x/crypto v0.23.0
+	golang.org/x/time v0.5.0
+)
+
+require (
+	github.com/cespare/xxhash/v2 v2.3.0 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/redis/go-redis/v9 v9.5.1 // indirect
 )
